+++ src/shared/components/event-start/components/task-assigned/task-assigned.component.spec.ts
@@ -60,7 +60,7 @@ describe('TaskRequirementComponent', () => {
   const judicialworker: Judicialworker = {
     idamId: '4321-4321-4321-4321',
     firstName: 'Test',
-    lastName: 'Judicial User',
+    lastName: 'Judicialworker',
     email: 'testuser@demoenv.com',
     location: null
   }
@@ -100,7 +100,7 @@ describe('TaskRequirementComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(TaskAssignedComponent);
     component = fixture.componentInstance;
-    spyOn(mockCaseworkerService, 'getCaseworkers').and.returnValue(of([caseworker]));
+    spyOn(mockCaseworkerService, 'getCaseworkers').and.returnValue(of([]));
     spyOn(mockJudicialworkerService, 'getJudicialworkers').and.returnValue(of([judicialworker]));
     component.task.assignee = '4321-4321-4321-4321';
     fixture.detectChanges();
@@ -110,5 +110,16 @@ describe('TaskRequirementComponent', () => {
     const heading: DebugElement = fixture.debugElement.query(By.css('.govuk-heading-m'))
     const headingHtml = heading.nativeElement as HTMLElement;
     expect(headingHtml.innerText).toBe('Task assignment required');
+    expect(mockCaseworkerService.getCaseworkers).toHaveBeenCalled();
+    expect(mockJudicialworkerService.getJudicialworkers).toHaveBeenCalled();
+    expect(component.assignedUserName).toEqual('Test Judicialworker');
+  });
+
+  it('should unsubscribe subscriptions', () => {
+    spyOn(component.caseworkerSubscription, 'unsubscribe').and.callThrough();
+    spyOn(component.judicialworkerSubscription, 'unsubscribe').and.callThrough();
+    component.ngOnDestroy();
+    expect(component.caseworkerSubscription.unsubscribe).toHaveBeenCalled();
+    expect(component.judicialworkerSubscription.unsubscribe).toHaveBeenCalled();
   });
 });