+++ RELEASE-NOTES.md
@@ -1,7 +1,7 @@
 ## RELEASE NOTES
 
-### Version 4.9.9-state-machine
-**Task event completion and state machine** Task event completion and state machine
+### Version 4.9.10-state-machine"
+**Task event completion and state machine**
 
 ### Version 4.3.6-hotfix-EUI-4404
 **EUI-4404** Fix rendering of read-only Collection field within editable Complex field